fixes #130
